FEAT: error_page parsing implemented

+++ inc/Configparsing/modules.hpp
@@ -104,7 +104,7 @@ struct WebservErrorLog {
 };
 
 struct	ErrorPage {
-	unsigned int	resCode;
+	unsigned long	resCode;
 	std::string		path;
 };
